refactor: Update Hero component to use Icon component for grid icon
This commit refactors the Hero component to use the Icon component for displaying the grid icon. Instead of directly using an img tag, the Hero component now utilizes the Icon component with a size of 25 and the "bounce" animation. This change improves code modularity and enhances the consistency of using the Icon component throughout the application.