feat: Add public and authentication routes
Add an array of routes that are accessible to the public and do not require authentication/login. Also, include an array of routes used for authentication, which will redirect users to the /feed route. Additionally, define the prefix for API authentication routes and set the default redirect path after logging in.